feat: Make PreviewBrowser a component, force mobile theme
